# blog/docs/swagger.yaml
openapi: 3.0.3
info:
  description: "This is server for blog-platform server."
  version: "1.0.11"
  title: "Swagger blog-platform"
servers:
  - url: http://localhost:8080
...
paths:
  /api/create-user:
    post:
      tags:
        - users
      summary: "Create New User"
      description: "Create New User"
      operationId: "CreateUser"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '405':
          description: Invalid input
  /api/user/{user_id}:
    get:
      tags:
        - users
      summary: Find user by ID
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          description: ID of user to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '404':
          description: user not found
    put:
      tags:
        - users
      summary: Updates a user in the store with form data
      description: ''
      operationId: updateUserWithForm
      parameters:
        - name: user_id
          in: path
          description: ID of user that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful user update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request

    delete:
      tags:
        - users
      summary: Deletes a user
      description: delete a user
      operationId: deleteUser
      parameters:
        - name: user_id
          in: path
          description: user id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: No Content
        '400':
          description: Invalid user value
  /api/users:
    get:
      tags:
        - users
      summary: list all users
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/from"
      responses:
        '200':
          description: list all users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: internal server error
...
  /api/post/{post_id}/add-comment:
    post:
      tags:
        - comments
      summary: "Create New Comment"
      description: "Create New Comment"
      operationId: "CreateComment"
      parameters:
        - name: post_id
          in: path
          description: id of the particular post
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateComment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateComment'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '405':
          description: Invalid input
  /api/post/{post_id}/comments/{comment_id}:
    get:
      tags:
        - comments
      summary: Find comment by ID
      description: Returns a single comment
      operationId: getCommentById
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: comment_id
          in: path
          description: comment id that need to be return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid ID supplied
        '404':
          description: tag not found
    put:
      tags:
        - comments
      summary: Updates a comment in the store with form data
      description: ''
      operationId: updateCommentWithForm
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: comment_id
          in: path
          description: ID of comment that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
          application/xml:
            schema:
              $ref: '#/components/schemas/CreateComment'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateComment'
      responses:
        '200':
          description: successful tag update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad Request

    delete:
      tags:
        - comments
      summary: Deletes a comment
      description: delete a comment
      operationId: deleteComment
      parameters:
        - name: post_id
          in: path
          description: ID of particular post
          required: true
          schema:
            type: integer
            format: int64
        - name: comment_id
          in: path
          description: comment id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: No Content
        '400':
          description: Invalid tag value
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: theUser
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
      xml:
        name: user
...
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: "classic"
        content:
          type: string
          example: "test blog"
        author:
          $ref: '#/components/schemas/User'
        tags_id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
        updated_at:
          type: string
          format: date-time
          example: 2022-08-14T19:33:16.428870284+05:30
    CreateComment:
      type: object
      properties:
        name:
          type: string
          example: "classic"
        body:
          type: string
          example: "comments"
      xml:
        name: tag
    CommentResponse:
      type: object
      properties:
        createdId:
          type: integer
          example: 1
        name:
          type: string
          example: "classic"
        comments:
          type: string
          example: "comments"
        post_id:
          type: integer
          example: 1
    Comment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: "classic"
        body:
          type: string
          example: "test"
        post_id:
          type: integer
          example: 1
        post:
          $ref: '#/components/schemas/Post'
  parameters:
    limit:
      name: limit
      in: query
      description: Number of records to return
      schema:
        type: integer
        format: int64
    to:
      name: to
      in: query
      description: first index of the page
      schema:
        type: integer
        format: int64
    from:
      name: from
      in: query
      description: last index of the previous page
      schema:
        type: integer
        format: int64